% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b06.R
\name{b06}
\alias{b06}
\alias{b06.sheet}
\alias{b0601_init}
\alias{b06_1.sbi}
\alias{b06_2.sbi}
\alias{b06_MP}
\alias{b06_2_De}
\alias{b06_1_De}
\alias{b0601_De}
\title{【別表六(一) 所得税額の控除に関する明細書】
HOB016_6.0.csv # 2021(令和3年4月1日以後終了事業年度又は連結事業年度分)
同上           # 2020(令和2年4月1日以後終了事業年度又は連結事業年度分)
HOB016_5.0.csv # 2019(平成31年4月1日以後終了事業年度又は連結事業年度分)
HOB016_4.0.csv # 2018(平成30年4月1日以後終了事業年度又は連結事業年度分)}
\usage{
b06.sheet(s = "2021-04-01")

b0601_init(n)

b06_1.sbi(sbi, o = 1)

b06_2.sbi(sbi, o = 1)

b06_MP(b06_1, b06_2, i06_pt, i06_ct)

b06_2_De(b06_2, MP)

b06_1_De(b06_1, MP)

b0601_De(b06_1, b06_2, i06_pt, i06_ct)
}
\arguments{
\item{s}{Settlement date}

\item{n}{format}

\item{sbi}{df}

\item{o}{順序}

\item{b06_1}{個別法}

\item{b06_2}{銘柄別簡便法}

\item{i06_pt}{前期}

\item{i06_ct}{当期}

\item{MP}{df}
}
\description{
法人税申告書別表等(明細記載を要する部分)の
標準フォーム等
(令和3年4月1日以後終了事業年度又は連結事業年度分)
https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho2_beppyo_03.htm
}
\details{
○ レコードの内容及び留意事項
【別表六（一） 所得税額の控除に関する明細書】
（令和３年４月１日以後終了事業年度分）
https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho1/7/HOB016.pdf

https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho1/7/HOB016.pdf
}
