% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b08.R
\name{b08}
\alias{b08}
\alias{b08.sheet}
\alias{b0801_init}
\alias{b08_4.sbi}
\alias{b08_4.B1m}
\alias{b08_4.A2m}
\alias{b08_4.ex}
\title{【別表八(一) 受取配当等の益金不算入に関する明細書】
HOB800_16.0.csv # 2021(令和3年4月1日以後終了事業年度又は連結事業年度分)
HOB800_15.0.csv # 2020（令和２年)
HOB800_14.0.csv # 2019(平成31年4月1日以後終了事業年度又は連結事業年度分)
HOB800_13.0.csv # 2018(平成30年4月1日以後終了事業年度又は連結事業年度分)
s_b08 <- "HOB810_5.0"　# バージョン番号(_5.0)}
\usage{
b08.sheet(s = "2021-04-01")

b0801_init(n = 4)

b08_4.sbi(sbi, o = 1)

b08_4.B1m(b08_4, p)

b08_4.A2m(b08_4, idou)

b08_4.ex(b08_4, i06_pt, i06_ct)
}
\arguments{
\item{s}{Settlement date}

\item{n}{format}

\item{sbi}{df}

\item{o}{順序}

\item{b08_4}{df}

\item{p}{異動明細（前期＆当期の有価証券内訳書）}

\item{idou}{前期＆当期の有価証券内訳書}

\item{i06_pt}{前期}

\item{i06_ct}{当期}
}
\description{
法人税申告書別表等(明細記載を要する部分)の
標準フォーム等
(令和3年4月1日以後終了事業年度又は連結事業年度分)
https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho2_beppyo_03.htm
}
\details{
○ レコードの内容及び留意事項
【別表八（一） 受取配当等の益金不算入に関する明細書】
（令和３年４月１日以後終了事業年度分）
https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho1/7/HOB800.pdf
}
