% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o_FD.R
\name{FD}
\alias{FD}
\alias{Write_eTax}
\alias{Read_eTax}
\alias{Read_MS}
\alias{Write_FA}
\alias{y4md_gymd}
\alias{gymd_y4md}
\alias{ascii_jp}
\title{Write_eTax}
\usage{
Write_eTax(df, s)

Read_eTax(d, f, e = ".csv", n = 0)

Read_MS(d, f, e = ".csv", n = 0)

Write_FA(df, f)

y4md_gymd(gymd)

gymd_y4md(y4md)

ascii_jp(jp)
}
\arguments{
\item{df}{dataframe}

\item{s}{sheet name}

\item{d}{directory}

\item{f}{file name without ext}

\item{e}{ext of f default(.csv)}

\item{n}{skip line of f default( 0 )}

\item{gymd}{df(元号 年 月 日)}

\item{y4md}{西暦}

\item{jp}{西暦}
}
\description{
df:入力データフレーム型（行列型でも動くと思う）
↓
s:出力シート名（MSエクセルのあれが必須です）

ソフトウェア開発業者の方へ
https://www.e-tax.nta.go.jp/shiyo/index.htm
}
