% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i06.R
\name{i06}
\alias{i06}
\alias{i06.sheet}
\alias{i06_init}
\alias{i06_sbi_idou}
\alias{i06_kabu_idou}
\alias{i06_sbi_As}
\alias{i06_merge}
\alias{i06_Book_value}
\alias{i06_Xs}
\alias{i06_lines_sum}
\alias{i06_ct}
\alias{i06_check_idou}
\alias{i06_idou}
\alias{i06_p}
\title{【⑥有価証券の内訳書】
HOI060_4.0.csv # 2021(令和3年4月1日以後終了事業年度又は連結事業年度分)
同上           # 2020(令和2年4月1日以後終了事業年度又は連結事業年度分)
同上           # 2019(平成31年4月1日以後終了事業年度又は連結事業年度分)
HOI060_3.0.csv # 2018(平成30年4月1日以後終了事業年度又は連結事業年度分)
（命名規約)バージョン番号(_4.0)までは必須}
\usage{
i06.sheet(s = "2021-04-01")

i06_init()

i06_sbi_idou(sbi)

i06_kabu_idou(kabu)

i06_sbi_As(sbi)

i06_merge(df1, df2)

i06_Book_value(i06_pt, i06_ct)

i06_Xs(Xs, i06_pt, i06_ct)

i06_lines_sum(i06_0)

i06_ct(i06_0, i06_1 = i06_init(), n = 1)

i06_check_idou(i06)

i06_idou(i06)

i06_p(i06_pt, i06_ct)
}
\arguments{
\item{s}{Settlement date}

\item{sbi}{dataframe}

\item{kabu}{dataframe}

\item{df1}{dataframe}

\item{df2}{dataframe}

\item{i06_pt}{前期}

\item{i06_ct}{当期}

\item{Xs}{株式交換}

\item{i06_0}{明細行}

\item{i06_1}{合計行}

\item{n}{2 損益行}

\item{i06}{dataframe}
}
\description{
勘定科目内訳明細書の標準フォーム等
(平成31年4月1日以後終了事業年度分)
https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho1/2/HOI060.pdf
}
\details{
(平成30年4月1日以後終了事業年度分)
https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho2_meisai_30.htm

○ レコードの内容及び留意事項(2-6)
【勘定科目内訳明細書（⑥有価証券の内訳書）】
（平成31年４月１日以後終了事業年度分）
https://www.e-tax.nta.go.jp/hojin/gimuka/csv_jyoho1/2/HOI060.pdf
}
