% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/office-package.R
\docType{package}
\name{office-package}
\alias{office}
\alias{office-package}
\title{office: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Naohito Takahashi \email{nt1969m@gmail.com}

}
\keyword{internal}
