% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_i01.R, R/bank_read.R, R/bank_tb.R
\name{read_bank}
\alias{read_bank}
\alias{bank_i01}
\alias{bank_tb}
\title{bank_i01}
\usage{
bank_i01(df = "", inf = "", col_i01 = "")

read_bank(
  d = getwd(),
  f = ".csv$",
  p = "^Fin",
  cp = "CP932",
  i01 = "",
  tax = 15,
  fuk = 2.1
)

bank_tb(df, inf, tax, fuk)
}
\arguments{
\item{df}{file}

\item{inf}{pf}

\item{col_i01}{i01_cols}

\item{d}{folder}

\item{f}{file}

\item{p}{pf}

\item{cp}{code page}

\item{i01}{cols}

\item{tax}{income tax}

\item{fuk}{special tax}
}
\description{
bank_i01

read_bank

bank_tb
}
